PR Review Resolved
- 추가 파일 생성
  -> Contents(구조체)
  -> SettingInfoView(View)

- Binding 사용
